Renaming imports of tsx files
